+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## Version 2.3.0
 
 - Added error message to indicate no column names are found
+- Added TRUNCATE function to table view
 
 ## Version 2.2.3